restructuring cuda directory
