+++ adaboost/cuda/core/cuda_data_structures.hpp
