+++ adaboost/cuda/core/cuda_data_structures_impl.cu
