+++ adaboost/cuda/core/cuda_data_structures_impl.hpp
@@ -2,8 +2,8 @@
 #define CUDA_ADABOOST_CORE_DATA_STRUCTURES_IMPL_HPP
 
 #include<adaboost/utils/utils.hpp>
-#include<adaboost/utils/cuda_wrappers.hpp>
-#include<adaboost/cuda/cuda_data_structures.hpp>
+#include<adaboost/cuda/utils/cuda_wrappers.hpp>
+#include<adaboost/cuda/core/cuda_data_structures.hpp>
 #include<cmath>
 #include<iostream>
