+++ adaboost/tests/test_cuda_data_structures.hpp
@@ -1,7 +1,7 @@
 #include<gtest/gtest.h>
 #include<string>
-#include<adaboost/cuda/cuda_data_structures.hpp>
-#include<adaboost/utils/cuda_wrappers.hpp>
+#include<adaboost/cuda/core/cuda_data_structures.hpp>
+#include<adaboost/cuda/utils/cuda_wrappers.hpp>
 #include<stdexcept>
 
 TEST(Cuda, VectorGPU)
